Remove IBranch interface from BranchModel
