+++ src/GitHub.Exports.Reactive/ViewModels/GitHubPane/IPullRequestCreationViewModel.cs
@@ -9,9 +9,9 @@ namespace GitHub.ViewModels.GitHubPane
 {
     public interface IPullRequestCreationViewModel : IPanePageViewModel
     {
-        IBranch SourceBranch { get; set; }
-        IBranch TargetBranch { get; set; }
-        IReadOnlyList<IBranch> Branches { get; }
+        BranchModel SourceBranch { get; set; }
+        BranchModel TargetBranch { get; set; }
+        IReadOnlyList<BranchModel> Branches { get; }
         ReactiveCommand<Unit, IPullRequestModel> CreatePullRequest { get; }
         ReactiveCommand<Unit, Unit> Cancel { get; }
         string PRTitle { get; set; }
